+++ src/types.ts
@@ -1,18 +1,17 @@
+import { Notif } from './notif'
+
 export type OPT = {
-    debug: boolean
+    dynamic: boolean
     initMsgId?: number
     interval: number
-    notifyFreq: number
     sendFile: boolean
     session: string
     silent: boolean
     tags: string[]
     to?: string
     token?: string
-    _end?: number
-    _exit?: boolean
+    _notif: Notif
     _start: number
-    _output: string[]
     _version: string
 }